Fix ubsan triggered in gdb-add-index.sh
When compiling with CFLAGS="-O0 -ggdb3 -fsanitize=undefined"
LDFLAGS="-fsanitize=undefined' and running the testsuite, we run into:
...
dwz.c:573:64: runtime error: left shift of 144 by 24 places cannot be \
  represented in type 'int'
FAIL: src/testsuite/dwz.tests/gdb-add-index.sh
...

The problem is here in buf_read_ule32:
...
static inline uint32_t
buf_read_ule32 (unsigned char *data)
{
  return data[0] | (data[1] << 8) | (data[2] << 16) | (data[3] << 24);
...

The data[3] is an unsigned char, which is promoted to int when used as shift
operand, which causes the shift to be a signed one, which can result in
undefined behaviour when shifting into the sign bit.

Fix this by casting to unsigned int.  Likewise in buf_read_ube32.

2021-03-15  Tom de Vries  <tdevries@suse.de>

	* dwz.c (buf_read_ule32, buf_read_ube32): Avoid undefined signed
	left-shift behaviour.

+++ dwz.c
@@ -570,13 +570,15 @@ buf_read_ube16 (unsigned char *data)
 static inline uint32_t
 buf_read_ule32 (unsigned char *data)
 {
-  return data[0] | (data[1] << 8) | (data[2] << 16) | (data[3] << 24);
+  return (data[0] | (data[1] << 8) | (data[2] << 16)
+	  | ((unsigned int)data[3] << 24));
 }
 
 static inline uint32_t
 buf_read_ube32 (unsigned char *data)
 {
-  return data[3] | (data[2] << 8) | (data[1] << 16) | (data[0] << 24);
+  return (data[3] | (data[2] << 8) | (data[1] << 16)
+	  | ((unsigned int)data[0] << 24));
 }
 
 static inline uint64_t
